refactor(상품): 상품서비스 전체 조회 수정 + 상품Dto 이름 수정
