+++ src/main/java/kitchenpos/products/application/ProductService.java
@@ -3,7 +3,7 @@ package kitchenpos.products.application;
 import kitchenpos.menus.application.MenuService;
 import kitchenpos.products.domain.ProductRepository;
 import kitchenpos.products.dto.ProductChangePriceRequest;
-import kitchenpos.products.dto.ProductResponse;
+import kitchenpos.products.dto.ProductRequest;
 import kitchenpos.products.infra.PurgomalumClient;
 import kitchenpos.products.tobe.domain.Product;
 import kitchenpos.products.tobe.domain.ProductPrice;
@@ -31,7 +31,7 @@ public class ProductService {
     }
 
     @Transactional
-    public Product create(final ProductResponse request) {
+    public Product create(final ProductRequest request) {
         Product product = request.toEntity(purgomalumClient);
         return productRepository.save(product);
     }
