+++ src/main/java/kitchenpos/products/dto/ProductResponse.java
@@ -1,15 +1,19 @@
 package kitchenpos.products.dto;
 
 import kitchenpos.products.tobe.domain.Product;
-import kitchenpos.products.infra.PurgomalumClient;
 
 import java.math.BigDecimal;
+import java.util.UUID;
 
 public class ProductResponse {
+    private UUID id;
+
     private String name;
+
     private BigDecimal price;
 
-    public ProductResponse(String name, BigDecimal price) {
+    public ProductResponse(UUID id, String name, BigDecimal price) {
+        this.id = id;
         this.name = name;
         this.price = price;
     }
@@ -22,7 +26,10 @@ public class ProductResponse {
         return price;
     }
 
-    public Product toEntity(PurgomalumClient profanityClient){
-        return Product.of(this.name, profanityClient, this.price);
+    public static ProductResponse fromEntity(Product product) {
+        return new ProductResponse(
+                product.getId()
+                , product.getNameValue()
+                , product.getPriceValue());
     }
 }
