+++ src/main/java/kitchenpos/products/ui/ProductRestController.java
@@ -2,9 +2,9 @@ package kitchenpos.products.ui;
 
 import kitchenpos.products.application.ProductService;
 import kitchenpos.products.dto.ProductChangePriceRequest;
-import kitchenpos.products.tobe.domain.Product;
-import kitchenpos.products.dto.ProductResponse;
 import kitchenpos.products.dto.ProductRequest;
+import kitchenpos.products.dto.ProductResponse;
+import kitchenpos.products.tobe.domain.Product;
 import org.springframework.http.ResponseEntity;
 import org.springframework.web.bind.annotation.*;
 
@@ -12,6 +12,7 @@ import javax.validation.Valid;
 import java.net.URI;
 import java.util.List;
 import java.util.UUID;
+import java.util.stream.Collectors;
 
 @RequestMapping("/api/products")
 @RestController
@@ -23,19 +24,24 @@ public class ProductRestController {
     }
 
     @PostMapping
-    public ResponseEntity<ProductRequest> create(@Valid @RequestBody final ProductResponse request) {
+    public ResponseEntity<ProductResponse> create(@Valid @RequestBody final ProductRequest request) {
         final Product response = productService.create(request);
         return ResponseEntity.created(URI.create("/api/products/" + response.getId()))
-            .body(ProductRequest.fromEntity(response));
+                .body(ProductResponse.fromEntity(response));
     }
 
     @PutMapping("/{productId}/price")
-    public ResponseEntity<Product> changePrice(@PathVariable final UUID productId, @Valid @RequestBody final ProductChangePriceRequest request) {
-        return ResponseEntity.ok(productService.changePrice(productId, request));
+    public ResponseEntity<ProductResponse> changePrice(@PathVariable final UUID productId, @Valid @RequestBody final ProductChangePriceRequest request) {
+        final Product response = productService.changePrice(productId, request);
+        return ResponseEntity.ok(ProductResponse.fromEntity(response));
     }
 
     @GetMapping
     public ResponseEntity<List<ProductResponse>> findAll() {
-        return ResponseEntity.ok(productService.findAll());
+        final List<ProductResponse> responses = productService.findAll()
+                .stream()
+                .map(ProductResponse::fromEntity)
+                .collect(Collectors.toUnmodifiableList());
+        return ResponseEntity.ok(responses);
     }
 }
